restructured train dir
