+++ pyproject.toml
@@ -30,7 +30,8 @@ dependencies = [
   "scipy==1.11.4",
   "seaborn==0.13.2",
   "ucimlrepo==0.0.3",
-  "xgboost==2.0.3"
+  "xgboost==2.0.3",
+  "mlflow>=3.5.1",
 ]
 
 [project.optional-dependencies]
