+++ src/training/artifacts/experiment_keys.csv
@@ -1,2 +1,2 @@
 0,1
-random-forest,237bf2538e134070a20f0129b5e25da5
+lightgbm,0bbb38ca299d49778d3b74ee8ac40b4f
